[#1] refac : rename base_response, base_request
