+++ src/core/applications/dtos/user_dto.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from src.core.applications.dtos.base import BaseRequest
+from src.core.applications.dtos.base_request import BaseRequest
 from src.core.domain.entities.user_entity import (
     CreateUserEntity,
     UpdateUserEntity,
