+++ src/core/applications/responses/user_response.py
@@ -2,7 +2,7 @@
 from typing import List, Union
 
 from src.core.applications.dtos.user_dto import UserDto
-from src.core.applications.responses.base import BaseResponse
+from src.core.applications.responses.base_response import BaseResponse
 
 
 class UserResponse(BaseResponse):
