+++ src/server/application/controllers/user_controller.py
@@ -7,7 +7,7 @@ from dependency_injector.wiring import Provide, inject
 from fastapi import APIRouter, Depends
 
 from src.core.applications.dtos.user_dto import CreateUserDto, UpdateUserDto
-from src.core.applications.responses.base import BaseResponse
+from src.core.applications.responses.base_response import BaseResponse
 from src.core.applications.responses.user_response import UserResponse
 from src.server.application.use_cases.user_use_case import UserUseCase
 from src.server.infrastructure.di.server_container import ServerContainer
